% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permtools.R
\docType{package}
\name{permtools-package}
\alias{permtools}
\alias{permtools-package}
\title{sceptre}
\description{
\code{permtools} is an R package providing tools for permutation testing.
}
\author{
\strong{Maintainer}: Tim Barry \email{tbarry@hsph.harvard.edu}

}
\keyword{package}
