% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_hypotheses.R
\name{run_bc_multiple_test_r}
\alias{run_bc_multiple_test_r}
\title{Run BC multiple testing procedure}
\usage{
run_bc_multiple_test_r(X, Y, h = 15L, alpha = 0.1)
}
\arguments{
\item{X}{an n by m matrix of treatments}

\item{Y}{an n by m matrix of responses}

\item{h}{the tuning parameter anytime-balid BC p-value}
}
\value{
a list containing (1) the 1-based indices of the hypotheses that are significant, (2) the number of permutations for each hypothesis, and (3) the p-value for each hypothesis
}
\description{
Runs the permutation test based on the anytime-valid BC p-value and BH correction.
}
\examples{
n <- 100L
m <- 5000L
X <- matrix(sample(x = c(0L, 1L), size = n * m, replace = TRUE), nrow = n, ncol = m)
under_null <- sample(c(TRUE, FALSE), size = m, prob = c(0.95, 0.05), replace = TRUE)
Y <- sapply(X = seq_len(m), FUN = function(i) {
 col <- X[,i]
 Y <- rnorm(n = n, mean = 2 + (if (under_null[i]) 0 else 2 * col), sd = 1)
})
h <- 15L; alpha <- 0.1
r <- sapply(seq(1, m), function(i) cor(X[,i], Y[,i]))
res <- run_bc_multiple_test_r(X, Y, alpha = alpha)
}
