% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_hypotheses_v2.R
\name{run_bc_multiple_test_r_v2}
\alias{run_bc_multiple_test_r_v2}
\title{Run BC multiple testing procedure}
\usage{
run_bc_multiple_test_r_v2(x, Y, h = 15L, alpha = 0.1)
}
\arguments{
\item{x}{an n by m matrix of treatments}

\item{Y}{an n by m matrix of responses}

\item{h}{the tuning parameter anytime-balid BC p-value}
}
\value{
a list containing (1) the 1-based indices of the hypotheses that are significant, (2) the number of permutations for each hypothesis, and (3) the p-value for each hypothesis
}
\description{
Runs the permutation test based on the anytime-valid BC p-value and BH correction.
}
\examples{
n <- 100L
m <- 1000L
x <- sample(x = c(0L, 1L), size = n, replace = TRUE)
under_null <- sample(c(TRUE, FALSE), size = m, prob = c(0.95, 0.05), replace = TRUE)
Y <- sapply(X = seq_len(m), FUN = function(i) {
  rnorm(n = n, mean = 2 + (if (under_null[i]) 0 else 2 * x), sd = 1)
})
h <- 15L; alpha <- 0.1
res_r <- run_bc_multiple_test_r_v2(x, Y, alpha = alpha)
mean(under_null[res_r |> filter(significant) |> pull(hypothesis_idx)])

# c++ version
Y_list <- apply(X = Y, MARGIN = 2, FUN = function(col) col, simplify = FALSE)
res <- run_bc_multiple_test(x, Y_list, h, alpha)
mean(under_null[which(res)])
}
