% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_single_hypothesis.R
\name{run_bc_perm_test_r}
\alias{run_bc_perm_test_r}
\title{Run binomial mixture permutation test}
\usage{
run_bc_perm_test_r(x, y, h = 15L)
}
\arguments{
\item{x}{a binary, integer treatment vector of 1s and 0s}

\item{y}{a numeric response vector}

\item{h}{tuning parameter; the number of losses to incur until stopping the algorithm}
}
\value{
a list containing (i) the p-value and (ii) the numer of permutations
}
\description{
Run binomial mixture permutation test
}
\examples{
set.seed(4)
n <- 100
y <- rnorm(100)
x <- y > median(y)
swap_idxs <- sample(x = seq_along(x), size = 35)
x[swap_idxs] <- !x[swap_idxs]
res <- run_bc_perm_test_r(x, y)
}
