% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_implementation.R
\name{run_binom_mixture_perm_test}
\alias{run_binom_mixture_perm_test}
\title{Run binomial mixture permutation test}
\usage{
run_binom_mixture_perm_test(x, y, target_alpha = 0.05)
}
\arguments{
\item{x}{a binary, integer treatment vector of 1s and 0s}

\item{y}{a numeric response vector}
}
\value{
a list containing (i) the p-value and (ii) the numer of permutations
}
\description{
Run binomial mixture permutation test
}
\examples{
set.seed(4)
n <- 100
y <- rnorm(100)
x <- y > median(y)
swap_idxs <- sample(x = seq_along(x), size = 20)
x[swap_idxs] <- !x[swap_idxs]
target_alpha <- 0.05
res <- run_binom_mixture_perm_test(x, y, target_alpha)
}
